Move HttpRequestImpl back to Node directory
